+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- GUIScope.FontStyle for scoped font style changes.
+
 ## [0.20.0] - 2025-09-12
 
 ### Added